Add class for our specific GPS-module (Ublox)

+++ project.latex.balloon/src/project/latex/balloon/sensor/AltimeterSensorController.java
@@ -11,5 +11,5 @@ package project.latex.balloon.sensor;
  * @author dgorst
  */
 public interface AltimeterSensorController {
-    float getHeight();
+    double getHeight();
 }
